refactor tests to get intellij to parse our unittests
