+++ tests/test_multi_asset.py
@@ -4,7 +4,7 @@
 import datetime
 import unittest
 
-import tests.transaction_utils
+import transaction_utils
 from yabc import basis
 from yabc import coinpool
 from yabc.transaction import Operation
@@ -16,25 +16,25 @@ class MultiAssetTest(unittest.TestCase):
     def setUp(self) -> None:
         self.start = datetime.datetime.now()
         self.one_day = datetime.timedelta(1)
-        self.purchase_bch = tests.transaction_utils.make_buy(
+        self.purchase_bch = transaction_utils.make_buy(
             quantity=1,
             fees=0,
             subtotal=100,
             date=self.start - 2 * self.one_day,
             symbol="BCH",
         )
-        self.purchase_btc = tests.transaction_utils.make_transaction(
+        self.purchase_btc = transaction_utils.make_transaction(
             Operation.BUY, 1, 0, TEN_K, date=self.start - self.one_day
         )
 
     def test_sell_matches_symbol(self):
         """
-        Sort of an integration test (because it tests several classes underneath the process_all api)
+        Sort of an integration test (because it several classes underneath the process_all api)
 
         Make sure that when we sell BTC, we match with a previous BTC buy (and not BCH or something else)
         :return:
         """
-        sell_btc_no_profit = tests.transaction_utils.make_transaction(
+        sell_btc_no_profit = transaction_utils.make_transaction(
             Operation.SELL, 1, 0, TEN_K, date=self.start + self.one_day
         )
         # Note that they don't need to be in order.
